+++ frontend/src/components/Chart/index.ts
@@ -1,0 +1,5 @@
+import LineChart from './LineChart';
+import PieChart from './PieChart';
+import Tooltip from './Tooltip';
+
+export { LineChart, PieChart, Tooltip };
